fix: implement coordinate-based ground truth alignment for model comparisons
Replace flawed ground truth cropping logic with robust coordinate-based alignment
that ensures reconstructed and ground truth images have identical dimensions for
accurate metric calculation and visualization.

Key improvements:
- Add centralized `align_for_evaluation()` function in cropping.py that uses
  reconstruction bounds (scan coordinates + stitch patch size) to crop ground truth
- Replace error-prone `crop_to_non_uniform_region_with_buffer()` approach with
  coordinate-driven alignment in compare_models.py
- Fix data type issues in PSNR calculation (cv2.PSNR requires matching dtypes)
- Ensure consistent M=20 stitch patch size parameter across baseline and comparison scripts

Fixes issue where ground truth was not displayed in comparison plots due to
shape mismatches between reconstructed (185×185) and ground truth (232×232) images.
Now all arrays are perfectly aligned for evaluation and visualization.

+++ ptycho/evaluation.py
@@ -152,8 +152,8 @@ def psnr(target, pred, normalize = True, shift = False):
     for phase inputs, assume that global shift has already been taken care off
     """
     import cv2
-    target = np.array(target)
-    pred = np.array(pred)
+    target = np.array(target, dtype=np.float32)
+    pred = np.array(pred, dtype=np.float32)
     if normalize:
         scale = np.mean(target) / np.mean(pred)
     else:
